Feature jwt cration (#16)
* feature-jwtcreation: jwt created when sign up new shopper

The sign up and login for shoppers its working good, with endpoints /api/auth for sign up and /api/login for login, the token conatins the username

to do: include Role in the token and do the login validation with the email, not with the username like is actually working

* feature-jwtcreation: login validation with the email

Now the login is working with searching by email

to do: include the Role in the Jwt

* Role added in the JWT

What? Now the jwt have the role in the claim

Why? Because we want to validate the permissions of each user

* Token working good with email

What? authentication working with token
why? Was not working good

* Protected employee creation

what? admin user can create employees
why? now just the admins can create new employees

* Update branch

* Logout working well

What? logout service implemented
Why? for users can logout from the account

* Fixed login bug with email